document-classification-extraction/requirements.txt
streamlit==1.27.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
pytesseract==0.3.10
pdf2image==1.16.3
pillow==10.0.0
spacy==3.6.1
python-docx==0.8.11
pdfplumber==0.9.0
nltk==3.8.1
matplotlib==3.7.2